Reogranize code utils
