+++ spyder/utils/tests/test_programs.py
@@ -2,21 +2,25 @@
 #
 # Copyright © Spyder Project Contributors
 # Licensed under the terms of the MIT License
-
 """Tests for programs.py"""
 
+# Standard library imports
 import os
 import os.path as osp
 import sys
 
+# Third party impors
 from flaky import flaky
 import pytest
 
-from spyder.utils.programs import (run_python_script_in_terminal,
-                                   is_python_interpreter,
+# Local imports
+from spyder.utils.programs import (check_version, find_program,
+                                   get_application_icon,
+                                   get_installed_applications, get_temp_dir,
+                                   is_module_installed, is_python_interpreter,
                                    is_python_interpreter_valid_name,
-                                   find_program, shell_split, check_version,
-                                   is_module_installed, get_temp_dir)
+                                   parse_linux_desktop_entry,
+                                   run_python_script_in_terminal, shell_split)
 
 
 if os.name == 'nt':
@@ -161,6 +165,31 @@ def test_get_temp_dir_ensure_dir_exists():
 
     assert os.path.exists(another_call)
     assert another_call == temp_dir
+
+
+def test_get_installed_apps_and_icons(qtbot):
+    apps = get_installed_applications()
+    assert apps
+    for app in apps:
+        fpath = apps[app]
+        icon = get_application_icon(fpath)
+        assert icon
+        assert osp.isdir(fpath) or osp.isfile(fpath)
+
+
+@pytest.mark.skipif(not sys.platform.startswith('linux'),
+                    reason="Test for linux only")
+def test_parse_linux_desktop_entry():
+    apps = get_installed_applications()
+    for app in apps:
+        fpath = apps[app]
+        data = parse_linux_desktop_entry(fpath)
+        assert data
+
+        for key in ['name', 'icon_path', 'hidden', 'exec', 'type', 'fpath']:
+            assert key in data
+
+        assert fpath == data['fpath']
 
 
 if __name__ == '__main__':
